casting size of std::list to unsigned to get rid of gcc461 warning

+++ TopQuarkAnalysis/TopKinFitter/plugins/TtSemiLepKinFitProducer.h
@@ -302,7 +302,7 @@ void TtSemiLepKinFitProducer<LeptonCollection>::produce(edm::Event& evt, const e
   // starting with the JetComb having the smallest chi2
   // -----------------------------------------------------
 
-  if( FitResultList.size() < 1 ) { // in case no fit results were stored in the list (all fits aborted)
+  if( (unsigned)FitResultList.size() < 1 ) { // in case no fit results were stored in the list (all fits aborted)
     pPartonsHadP->push_back( fitter->fittedHadP()     );
     pPartonsHadQ->push_back( fitter->fittedHadQ()     );
     pPartonsHadB->push_back( fitter->fittedHadB()     );
